# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for service mini.note resource.",
        "title": "Mini Note RESTful API",
        "contact": {
            "name": "API Support"
        },
        "version": "0.1"
    },
    "host": "localhost:9077",
    "basePath": "/api/v1",
    "paths": {
...
        "/notes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "View notes list",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Notes list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Number page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Number per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note items list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/UniversalDTO"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Notes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create note record",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Note record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note body",
                        "name": "note",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note category",
                        "name": "category_id",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "draft",
                            "public",
                            "archive"
                        ],
                        "type": "string",
                        "description": "Note status",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create note item",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            }
        },
        "/notes/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search note records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Note search",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "note search by q",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search note items list",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "View note element",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Note element",
                "parameters": [
                    {
                        "type": "string",
                        "description": "note id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note item",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/UniversalDTO"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/Note"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update note record",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Note record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "note id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note body",
                        "name": "note",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "draft",
                            "public",
                            "archive"
                        ],
                        "type": "string",
                        "description": "Note status",
                        "name": "status",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update note item",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete note record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Note delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "note id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete note record",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "type": "string",
                    "example": "Note not found."
                }
            }
        },
...
        "Note": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "Notes": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Note"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "UniversalDTO": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/Error"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}